Automatically generate the probe API with CMake
