+++ ebpf/src/bcc_probe_generator.cpp
@@ -15,6 +15,7 @@
  */
 
 #include "bcc_probe_generator.h"
+#include <bcc_probe_api.h>
 
 #include <fstream>
 #include <iostream>
@@ -28,163 +29,6 @@ namespace boostfs = boost::filesystem;
 
 namespace trailofbits {
 namespace {
-static const std::string kManagedProbeHeader = R"PROBE_HEADER(
-#include <linux/fs.h>
-#include <linux/sched.h>
-#include <uapi/linux/ptrace.h>
-
-#define EVENT_MAP_SIZE 1000000
-
-#define VARARGS_TERMINATOR 0xFFFF0000FFFF0000ULL
-#define VARARGS_TRUNCATION 0x0011001100110011ULL
-
-#define BASE_EVENT_TYPE 0x1122334455660000ULL
-
-BPF_PERF_OUTPUT(events);
-BPF_PERCPU_ARRAY(perf_event_data, u64, EVENT_MAP_SIZE);
-BPF_PERCPU_ARRAY(perf_cpu_index, u64, 1);
-
-#define INCREMENT_EVENT_DATA_INDEX_BY(idx, amount) \
-  idx = ((idx + amount) & 0x00FFFFFFUL) % EVENT_MAP_SIZE
-
-#define INCREMENT_EVENT_DATA_INDEX(idx) \
-  INCREMENT_EVENT_DATA_INDEX_BY(idx, 1)
-
-/// Saves the generic event header into the per-cpu map, returning the
-/// initial index
-static int saveEventHeader(u64 event_identifier,
-                           u64 syscall_number,
-                           bool save_exit_code,
-                           int exit_code) {
-  int index_key = 0U;
-  u64 initial_slot = 0U;
-  u64* index_ptr = perf_cpu_index.lookup_or_init(&index_key, &initial_slot);
-  int event_index = (index_ptr != NULL ? *index_ptr : initial_slot);
-
-  int index = event_index;
-  perf_event_data.update(&index, &event_identifier);
-  INCREMENT_EVENT_DATA_INDEX(index);
-
-  perf_event_data.update(&index, &syscall_number);
-  INCREMENT_EVENT_DATA_INDEX(index);
-
-  u64 field = bpf_ktime_get_ns();
-  perf_event_data.update(&index, &field);
-  INCREMENT_EVENT_DATA_INDEX(index);
-
-  field = bpf_get_current_pid_tgid();
-  perf_event_data.update(&index, &field);
-  INCREMENT_EVENT_DATA_INDEX(index);
-
-  field = bpf_get_current_uid_gid();
-  perf_event_data.update(&index, &field);
-  INCREMENT_EVENT_DATA_INDEX(index);
-
-  if (save_exit_code == true) {
-    field = (u64)exit_code;
-    perf_event_data.update(&index, &field);
-    INCREMENT_EVENT_DATA_INDEX(index);
-  }
-
-  initial_slot = index; // re-use the same var to avoid wasting stack space
-  perf_cpu_index.update(&index_key, &initial_slot);
-
-  return event_index;
-}
-
-/// Saves the given string into the per-cpu map
-static int saveStringBuffer(const char* buffer) {
-  int index_key = 0U;
-  u64 initial_slot = 0U;
-  u64* index_ptr = perf_cpu_index.lookup_or_init(&index_key, &initial_slot);
-  int index = (index_ptr != NULL ? *index_ptr : initial_slot);
-
-#pragma unroll
-  for (int i = 0; i < STRING_BUFFER_SIZE / 8; i++) {
-    perf_event_data.update(&index, (u64*)&buffer[i * 8]);
-    INCREMENT_EVENT_DATA_INDEX(index);
-  }
-
-  initial_slot = index; // re-use the same var to avoid wasting stack space
-  perf_cpu_index.update(&index_key, &initial_slot);
-
-  return 0;
-}
-
-/// Saves the string pointed to by the given address into the per-cpu map
-static bool saveString(char* buffer, const char* address) {
-  if (address == NULL) {
-    return false;
-  }
-
-  bpf_probe_read(buffer, STRING_BUFFER_SIZE, address);
-  saveStringBuffer(buffer);
-
-  return true;
-}
-
-#define saveByteArray saveString
-
-/// Saves the truncation identifier into the per-cpu map; used for varargs
-/// functions likes execve
-static int emitVarargsTerminator(bool truncated) {
-  int index_key = 0U;
-  u64 initial_slot = 0U;
-  u64* index_ptr = perf_cpu_index.lookup_or_init(&index_key, &initial_slot);
-  int index = (index_ptr != NULL ? *index_ptr : initial_slot);
-
-  u64 terminator = truncated == true ? VARARGS_TRUNCATION : VARARGS_TERMINATOR;
-  perf_event_data.update(&index, &terminator);
-  INCREMENT_EVENT_DATA_INDEX(index);
-
-  initial_slot = index; // re-use the same var to avoid wasting stack space
-  perf_cpu_index.update(&index_key, &initial_slot);
-
-  return 0;
-}
-
-/// Saves the given value to the per-cpu buffer; only use after the header
-/// has been sent
-#define saveSignedInteger saveEventValue
-#define saveUnsignedInteger saveEventValue
-static int saveEventValue(u64 value) {
-  int index_key = 0U;
-  u64 initial_slot = 0U;
-  u64* index_ptr = perf_cpu_index.lookup_or_init(&index_key, &initial_slot);
-  int index = (index_ptr != NULL ? *index_ptr : initial_slot);
-
-  perf_event_data.update(&index, &value);
-  INCREMENT_EVENT_DATA_INDEX(index);
-
-  initial_slot = index; // re-use the same var to avoid wasting stack space
-  perf_cpu_index.update(&index_key, &initial_slot);
-
-  return 0;
-}
-
-static int saveStringList(char *buffer, const char *const * string_list) {
-  const char* argument_ptr = NULL;
-
-#pragma unroll
-  for (int i = 1; i < STRING_LIST_SIZE; i++) {
-    bpf_probe_read(&argument_ptr, sizeof(argument_ptr), &string_list[i]);
-    if (saveString(buffer, argument_ptr) == false) {
-      goto emit_terminator;
-    }
-  }
-
-  goto emit_truncation;
-
-emit_truncation:
-  emitVarargsTerminator(true);
-  return 0;
-
-emit_terminator:
-  emitVarargsTerminator(false);
-  return 0;
-}
-)PROBE_HEADER";
-
 std::ostream& operator<<(
     std::ostream& stream,
     ManagedTracepointDescriptor::Parameter::Type tracepoint_param_type) {
@@ -306,10 +150,6 @@ std::string generateTracepointHandler(
   buffer << "  return 0;\n";
   buffer << "}\n\n";
 
-  std::cout << "\n\n======================\n"
-            << buffer.str() << "\n=========================\n\n"
-            << std::endl;
-
   return buffer.str();
 }
 
@@ -369,7 +209,7 @@ osquery::Status generateManagedTracepointProbeSource(
   buffer << "#define STRING_BUFFER_SIZE " << desc.string_buffer_size << "\n";
   buffer << "#define STRING_LIST_SIZE " << desc.string_list_size << "\n\n";
 
-  buffer << kManagedProbeHeader << "\n";
+  buffer << kBccProbeAPI << "\n";
 
   std::size_t tracepoint_id = 0U;
   for (const auto& tracepoint_desc : desc.tracepoint_list) {
@@ -382,6 +222,7 @@ osquery::Status generateManagedTracepointProbeSource(
   }
 
   probe_source_code = buffer.str();
+  std::cout << "\n\n\n\n" << probe_source_code << "\n\n\n\n" << std::endl;
   return osquery::Status(0);
 }
 } // namespace
